requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.3
pydantic>=2.4.0
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
mcp>=1.0.0
kaleido>=0.2.1
python-dotenv>=1.0.0
